+++ HISTORY.rst
@@ -10,6 +10,7 @@ aiofastnet Release History
 
 * Fixed Protocol.get_buffer not requiring PyBUF_WRITABLE from user buffer
 * Cleaned up nogil usage, harden logic against multithreading misuse by user code.
+* Fixed potential double-free error in SSLObject.__init__
 
 0.7.0
 ------------------