+++ HISTORY.rst
@@ -11,6 +11,7 @@ aiofastnet Release History
 * Fixed Protocol.get_buffer not requiring PyBUF_WRITABLE from user buffer
 * Cleaned up nogil usage, harden logic against multithreading misuse by user code.
 * Fixed potential double-free error in SSLObject.__init__
+* Fixed potential hangups on invalid ssl_handshake_timeout and ssl_shutdown_timeout values
 
 0.7.0
 ------------------